classical/aml-cli-v2/data-science/src/requirements/train-requirements.txt
numpy==1.22.0
azure-ai-ml
mlflow
azureml-mlflow
mltable